Added Wall Follow and FinalAutonomous
Wall Follow is the code provided by modern robotics which allows a robot
to follow a wall from a certain distance. The final autonomous is
comprised of the code from EncoderTest. ODSTest is basically EncoderTest
with if statements pertaining to detected light values inside of the if
statement that dictates do nothing until the encoderDrive method is
complete.